Convert int, to double, we can go from year -5000bc to year 2015 now

+++ src/calculatricegraphique/ADD.java
@@ -16,7 +16,7 @@ public class ADD extends Binary {
         right = e2;
     }
 
-    public int eval() {
+    public double eval() {
         return left.eval() + right.eval();
     }
 }
